+++ code/cgame/cg_event.c
@@ -254,9 +254,7 @@ static void CG_Obituary(entityState_t *ent) {
 	if (attacker == cg.snap->ps.clientNum) {
 		char *s;
 
-		if (!CG_IsATeamGametype(cgs.gametype) && !CG_UsesTeamFlags(cgs.gametype) &&
-				!CG_UsesTheWhiteFlag(cgs.gametype) && !CG_IsARoundBasedGametype(cgs.gametype) &&
-				cgs.gametype != GT_DOUBLE_D && cgs.gametype != GT_DOMINATION) {
+		if (cgs.gametype == GT_FFA || cgs.gametype == GT_TOURNAMENT) {
 			s = va("You fragged %s\n%s place with %i", targetName,
 					CG_PlaceString(cg.snap->ps.persistant[PERS_RANK] + 1),
 					cg.snap->ps.persistant[PERS_SCORE]);