+++ code/game/g_combat.c
@@ -133,7 +133,7 @@ void TossClientItems( gentity_t *self )
 		}
 	}
 
-	if (g_instantgib.integer || g_weaponArena.integer || g_gametype.integer == GT_CTF_ELIMINATION || g_elimination_allgametypes.integer) {
+	if (g_gametype.integer == GT_CTF_ELIMINATION || G_IsANoPickupsMode()) {
 		//Nothing!
 	}
 	else if ( weapon > WP_MACHINEGUN && weapon != WP_GRAPPLING_HOOK &&